+++ zcash_keys/CHANGELOG.md
@@ -7,7 +7,9 @@ and this library adheres to Rust's notion of
 ## [Unreleased]
 
 ### Added
-- `zcash_keys::keys::HdSeedFingerprint`
+- `zcash_keys::keys`:
+  - `HdSeedFingerprint`
+  - `UnifiedSpendingKey::{has_orchard, has_sapling, has_transparent}`
 - `zcash_keys::address::Address::has_receiver`
 - `impl Display for zcash_keys::keys::AddressGenerationError`
 - `impl std::error::Error for zcash_keys::keys::AddressGenerationError`